use cartopys builtin features
